Add quick fix for removing dollar prefix
